+++ README.md
@@ -24,15 +24,20 @@
 
 ```
 minute_factory/
-├── processing/           # 핵심 처리 로직 모음
-│   ├── api_helpers.py    # API 키 로드 헬퍼
-│   ├── audio_input.py    # (신규) 오디오 입력 UI/로직 모듈
-│   ├── audio_processing.py # 화자 분리 및 STT
-│   ├── chatbot_crag.py   # (신규) CRAG 챗봇 모듈
-│   ├── config.py         # 프로젝트 설정
-│   ├── file_io.py        # 결과 파일 저장
-│   ├── llm_handler.py    # GPT/Gemini 로직 처리
-│   └── pipeline.py       # 전체 처리 과정 총괄
+├── main_app.py           # Gradio 앱 실행 스크립트
+├── ui/                     # UI 관련 모듈
+│   ├── app_layout.py       # 전체 UI 레이아웃 및 콜백
+│   └── input_handlers.py   # 오디오 입력(업로드, 녹음) 처리
+├── core_processing/      # 핵심 기능 처리 모듈
+│   ├── main_pipeline.py    # 오디오 처리 전체 파이프라인
+│   ├── stt_diarization.py  # STT 및 화자 분리
+│   └── llm_tasks.py        # LLM 기반 교정 및 요약
+├── chatbot/                # 챗봇 관련 모듈
+│   └── crag_logic.py       # CRAG 챗봇 로직
+├── utils/                  # 공통 유틸리티 모듈
+│   ├── api_keys.py         # API 키 로드
+│   ├── config.py           # 전역 설정
+│   └── file_manager.py     # 파일 저장/관리
 ├── chroma_db/            # (자동생성) ChromaDB 벡터 저장소
 ├── data/                 # (자동생성) 입력 오디오 파일 저장
 ├── results/              # (자동생성) 결과물 저장
