+++ 작업기록.md
@@ -8,7 +8,7 @@
 
 ## 🎯 프로젝트 목표
 
-유수현, 성정현, 손용훈 3인의 파이썬 코드를 병합하여, 음성 파일을 입력받아 STT, 화자 분리, LLM 기반 교정 및 요약을 수행하는 단일 웹 애플리케이션(`Minute Factory`)을 개발. 최종적으로 회의록 기반의 Q&A가 가능한 챗봇 기능까지 통합한다.
+유수현, 성정현, 손용훈, 심세아 4인의 파이썬 코드를 병합하여, 음성 파일을 입력받아 STT, 화자 분리, LLM 기반 교정 및 요약을 수행하는 단일 웹 애플리케이션(`Minute Factory`)을 개발. 최종적으로 회의록 기반의 Q&A가 가능한 챗봇 기능까지 통합한다.
 
 ---
 
@@ -40,6 +40,14 @@
     - **챗봇 기능 오류 해결**: 웹 검색 결과 처리 시 발생하던 `TypeError`를 해결하여 챗봇 기능의 안정성 확보.
     - **라이브러리 경고 해결**: 최신 버전에 맞게 `Tavily`, `Chroma` 등 LangChain 관련 라이브러리의 `import` 구문과 사용법을 수정하여 모든 Deprecation 경고를 제거.
     - **UI/UX 개선**: Gradio 챗봇의 UI 포맷을 최신 `messages` 타입으로 업데이트하고, 관련 콜백 함수 로직을 수정.
+
+### 프로젝트 구조 리팩토링 (협업 효율성 증대)
+
+- **목표**: 팀원(유수현:UI, 성정현/손용훈:핵심처리, 심세아:챗봇)이 각자 담당 코드에 집중할 수 있도록 프로젝트 구조를 역할 기반으로 재구성.
+- **주요 변경 사항**:
+    - `processing` 폴더를 `ui`, `core_processing`, `chatbot`, `utils`의 4개 역할별 폴더로 분리.
+    - `main_app.py`에서 UI 생성 로직을 `ui/app_layout.py`로 이전하여, `main_app.py`는 앱 실행 역할만 담당하도록 간소화.
+    - 파일 이동 및 이름 변경에 따라 프로젝트 전체의 `import` 구문을 수정하여 새로운 구조에 맞게 업데이트.
 
 ---
